Proof of concept for test refactoring.
